+++ R/lotrrs.R
@@ -44,42 +44,7 @@
 #' @export 
 #'
 #' @examples
-#' if (interactive()) {
-#' # Use 'extdata' from the {flowWorkspaceData} package
-#'   flowDataPath <- system.file("extdata", package = "flowWorkspaceData")
-#'   fcsFiles <- list.files(pattern = "CytoTrol", flowDataPath, full = TRUE)
-#'   ncfs  <- ncdfFlow::read.ncdfFlowSet(fcsFiles)
-#'   fr1 <- ncfs[[1]]
-#'   fr2 <- ncfs[[2]]
-#' 
-#' # Comparison of two samples at two time points (two conditions) "g1" and "g2"
-#' ## (Create a random binary variable for "g2")
-#' ## One gate (two markers) "CD4", "CD38"
-#' ## Log10 Transformation for both markers
-#' ## Remove cells with NA and Inf values
-#' 
-#' # First sample
-#'   obs_dat1 <- data.frame("id" = seq(1, nrow(fr1@exprs), 1),
-#'                          "g1" = rep(1, nrow(fr1@exprs)),
-#'                          "g2" = stats::rbinom(nrow(fr1@exprs), 1, 0.5),
-#'                          "log10_CD4" = log(fr1@exprs[ , 5], 10),
-#'                          "log10_CD38" = log(fr1@exprs[ , 6], 10))
-#' # Second sample
-#'   obs_dat2 <- data.frame("id" = seq(1, nrow(fr2@exprs), 1),
-#'                          "g1" = rep(2, nrow(fr2@exprs)),
-#'                          "g2" = stats::rbinom(nrow(fr2@exprs), 1, 0.5),
-#'                          "log10_CD4" = log(fr2@exprs[ , 5], 10),
-#'                          "log10_CD38" = log(fr2@exprs[ , 6], 10))
-#' # Full set
-#'   obs_dat <- rbind(obs_dat1, obs_dat2)
-#'   obs_dat <- obs_dat[complete.cases(obs_dat), ] # remove NAs
-#'   obs_dat <- obs_dat[is.finite(rowSums(obs_dat)), ] # remove Infs
-#'   obs_dat$g1 <- as.factor(obs_dat$g1) # set "g1" as binary factor
-#'   obs_dat$g2 <- as.factor(obs_dat$g2) # set "g2" as binary factor
-#' 
-#' # Run lotrrs() function
-#'   test_lotrrs <- lotrrs(dat = obs_dat, p_correct = "none")
-#' }   
+#' test_lotrrs <- lotrrs(dat = randCyto)
 #' 
 lotrrs <- function(dat, 
                    alpha = 0.05,