+++ NEWS.md
@@ -1,5 +1,7 @@
 # gateR (development version)
 
-# gateR v0.1.3.9000
-
-* 
+# gateR v0.1.3
+  * Removed 'ncdFlow', 'flowWorkspaceData', and 'knitr' from Suggests
+  * Created a random data set 'randCyto' and all documentation
+  * Updated examples and testthat to use 'randCyto' data
+  * Updated vignette with clearer language